requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.3
python-jose==3.3.0
passlib==1.7.4
python-magic==0.4.27
pydantic==2.4.2 